Improve review-values.html build and output generation implementations 2
